dashboard/requirements.txt
streamlit==1.32.2
plotly==5.20.0
pandas==2.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
python-dotenv==1.0.1
